Manually adding recommend dll
I didn't find any other way to do it, as VisualStudio won't let me add external reference, and Unity doesn't provide a way to do it.
Info for this change is described in "known issues", here:  https://firebase.google.com/docs/unity/setup

.NET compatibility when using Unity 2017.x and later

Firebase supports .NET 4.x as an experimental build option in Unity 2017 and later. Firebase plugins use components of the Parse SDK to provide some .NET 4.x classes in earlier versions of .NET.

Therefore, the Firebase Unity SDK version 5.4.0 and later provide plugins that are compatible with either .NET 3.x or .NET 4.x in dotnet3 and dotnet4 directories of the Firebase Unity SDK.

If you import a Firebase plugin that is incompatible with the .NET version enabled in your project, you'll see compile errors from some types in the .NET framework that are implemented by the Parse SDK.

To resolve the compilation error, if you're using .NET 3.x:

    Remove or disable the following DLLs for all platforms:
        Parse/Plugins/dotNet45/Unity.Compat.dll
        Parse/Plugins/dotNet45/Unity.Tasks.dll
    Enable the following DLLs for all platforms:
        Parse/Plugins/Unity.Compat.dll
        Parse/Plugins/Unity.Tasks.dll