#1135 Rich text for UI elements (#1437)
New text rendering system for UI elements based on Unity TextMesh Pro plugin. For details on available tags see here:
http://digitalnativestudios.com/textmeshpro/docs/rich-text/
Also supports "justified" alignment, which is not mentioned in the rich text docs.

Technically it also includes some improvements for translators. It would be a pain to translate the rich UI format with the current quote escape mechanism, so there's a new format that uses "|||" characters instead for strings that contain a quote, but does not contain "|||".